+++ CHANGELOG.md
@@ -22,6 +22,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
   
 ### 🔄 Changed
 - Refactored ASI user to journey to use numpy arrays only
+- Refactored stratified PPI to accept numpy arrays as input
+- Restructured "User guide" into three pages "Evaluation Workflow", "Samplers" and "Estimators"
+- Logo
 - Decoupled Stratified PPI from PPI
 - `CONTRIBUTING.md` now explains how to setup pre-commits
 - Image in the user guide explaining the control sample